Apply /simplify findings: extract otlp_proto module, unify project injectors
- Move the protobuf wire-format encoder to core/otlp_proto.py (core/ is
  organized by concern; the encoder is a self-contained codec) with its
  tests in tests/core/test_otlp_proto.py.
- Collapse _inject_arize_project_name and
  _inject_openinference_project_resource_attr into one parameterized
  _inject_project_attr, reusing _to_otlp_attr_value for the AnyValue.
- Hoist function-local copy/struct/base64 imports to module level.
- Dedupe repeated-KeyValue encoding behind _pb_key_values (four sites).
- Drop unreachable coercion fallbacks in _pb_any_value (unknown or
  malformed values now encode as an empty AnyValue) and flatten the
  scope/status guards.
- Simplify the send_span Phoenix test to the payload-equality assertion;
  wire-level decoding assertions live with the encoder tests.
