+++ tests/core/test_common.py
@@ -15,7 +15,6 @@ from core.common import (
     StateManager,
     _attrs_to_otlp,
     _inject_openinference_project_resource_attr,
-    _otlp_json_to_protobuf,
     _resolve_kind,
     _to_otlp_attr_value,
     build_multi_span,
@@ -30,6 +29,7 @@ from core.common import (
     restore_stderr_from_log_file,
     send_span,
 )
+from core.otlp_proto import otlp_json_to_protobuf
 
 # ── Logging tests ──────────────────────────────────────────────────────────
 
@@ -880,157 +880,6 @@ class TestBuildMultiSpan:
         assert scope["name"] == "override-scope"
 
 
-# ── Minimal protobuf wire-format decoder (test helper) ────────────────────
-
-
-def _pb_read_varint(data, i):
-    result = shift = 0
-    while True:
-        byte = data[i]
-        result |= (byte & 0x7F) << shift
-        i += 1
-        if not byte & 0x80:
-            return result, i
-        shift += 7
-
-
-def _pb_decode(data):
-    """Decode a protobuf message into {field: [values]} (varint ints, raw bytes otherwise)."""
-    fields = {}
-    i = 0
-    while i < len(data):
-        tag, i = _pb_read_varint(data, i)
-        field, wire_type = tag >> 3, tag & 7
-        if wire_type == 0:
-            value, i = _pb_read_varint(data, i)
-        elif wire_type == 1:
-            value = data[i : i + 8]
-            i += 8
-        elif wire_type == 2:
-            length, i = _pb_read_varint(data, i)
-            value = data[i : i + length]
-            i += length
-        else:
-            raise ValueError(f"unsupported wire type {wire_type}")
-        fields.setdefault(field, []).append(value)
-    return fields
-
-
-def _pb_fixed64_int(raw):
-    import struct
-
-    return struct.unpack("<Q", raw)[0]
-
-
-# ── OTLP protobuf encoding tests ──────────────────────────────────────────
-
-
-class TestOtlpProtobufEncoding:
-    def test_encodes_span_fields(self):
-        payload = {
-            "resourceSpans": [
-                {
-                    "resource": {"attributes": [{"key": "service.name", "value": {"stringValue": "svc"}}]},
-                    "scopeSpans": [
-                        {
-                            "scope": {"name": "scope"},
-                            "spans": [
-                                {
-                                    "traceId": "0123456789abcdef0123456789abcdef",
-                                    "spanId": "abcdef1234567890",
-                                    "parentSpanId": "1122334455667788",
-                                    "name": "tool-call",
-                                    "kind": 1,
-                                    "startTimeUnixNano": "1000000000",
-                                    "endTimeUnixNano": "1500000000",
-                                    "attributes": [
-                                        {"key": "openinference.span.kind", "value": {"stringValue": "TOOL"}},
-                                        {"key": "count", "value": {"intValue": "3"}},
-                                        {"key": "ok", "value": {"boolValue": True}},
-                                        {"key": "score", "value": {"doubleValue": 0.5}},
-                                    ],
-                                    "events": [
-                                        {
-                                            "name": "exception",
-                                            "timeUnixNano": "1250000000",
-                                            "attributes": [{"key": "message", "value": {"stringValue": "boom"}}],
-                                        }
-                                    ],
-                                    "status": {"code": 2, "message": "failed"},
-                                }
-                            ],
-                        }
-                    ],
-                }
-            ]
-        }
-
-        encoded = _otlp_json_to_protobuf(payload)
-
-        # ExportTraceServiceRequest.resource_spans[0]
-        rs = _pb_decode(_pb_decode(encoded)[1][0])
-        # Resource.attributes[0] is service.name
-        resource_attr = _pb_decode(_pb_decode(rs[1][0])[1][0])
-        assert resource_attr[1][0] == b"service.name"
-        assert _pb_decode(resource_attr[2][0])[1][0] == b"svc"
-
-        ss = _pb_decode(rs[2][0])
-        assert _pb_decode(ss[1][0])[1][0] == b"scope"
-
-        span = _pb_decode(ss[2][0])
-        assert span[1][0] == bytes.fromhex("0123456789abcdef0123456789abcdef")
-        assert span[2][0] == bytes.fromhex("abcdef1234567890")
-        assert span[4][0] == bytes.fromhex("1122334455667788")
-        assert span[5][0] == b"tool-call"
-        assert span[6][0] == 1
-        assert _pb_fixed64_int(span[7][0]) == 1_000_000_000
-        assert _pb_fixed64_int(span[8][0]) == 1_500_000_000
-
-        attrs = {}
-        for raw in span[9]:
-            kv = _pb_decode(raw)
-            attrs[kv[1][0]] = _pb_decode(kv[2][0])
-        assert attrs[b"openinference.span.kind"][1][0] == b"TOOL"
-        assert attrs[b"count"][3][0] == 3
-        assert attrs[b"ok"][2][0] == 1
-        import struct
-
-        assert struct.unpack("<d", attrs[b"score"][4][0])[0] == 0.5
-
-        event = _pb_decode(span[11][0])
-        assert _pb_fixed64_int(event[1][0]) == 1_250_000_000
-        assert event[2][0] == b"exception"
-        event_attr = _pb_decode(event[3][0])
-        assert event_attr[1][0] == b"message"
-
-        status = _pb_decode(span[15][0])
-        assert status[2][0] == b"failed"
-        assert status[3][0] == 2
-
-    def test_rejects_missing_span_timestamp(self):
-        payload = {
-            "resourceSpans": [
-                {
-                    "scopeSpans": [
-                        {
-                            "spans": [
-                                {
-                                    "traceId": "t" * 32,
-                                    "spanId": "s" * 16,
-                                    "name": "missing-time",
-                                    "endTimeUnixNano": "2000000000",
-                                }
-                            ]
-                        }
-                    ]
-                }
-            ]
-        }
-
-        with pytest.raises(ValueError, match="Invalid Unix nanosecond timestamp"):
-            _otlp_json_to_protobuf(payload)
-
-
 # ── Phoenix project-name injection tests ──────────────────────────────────
 
 
@@ -1248,16 +1097,9 @@ class TestSendSpan:
         assert req.method == "POST"
         # Body is the OTLP payload with openinference.project.name appended to
         # the resource attributes for project routing, encoded as protobuf.
+        # (Encoding correctness itself is covered in test_otlp_proto.py.)
         expected_payload = _inject_openinference_project_resource_attr(self._SAMPLE_SPAN, "my-project")
-        assert req.data == _otlp_json_to_protobuf(expected_payload)
-        # Sanity-check the encoded resource attributes include the project name.
-        rs = _pb_decode(_pb_decode(req.data)[1][0])
-        resource_attrs = {}
-        for raw in _pb_decode(rs[1][0])[1]:
-            kv = _pb_decode(raw)
-            resource_attrs[kv[1][0]] = _pb_decode(kv[2][0])[1][0]
-        assert resource_attrs[b"openinference.project.name"] == b"my-project"
-        assert resource_attrs[b"service.name"] == b"test-service"
+        assert req.data == otlp_json_to_protobuf(expected_payload)
 
     @mock.patch("core.common.resolve_backend")
     @mock.patch("core.common.urllib.request.urlopen")
